requirements.txt
openai==0.27.8
python-dotenv==1.0.0
slack_bolt==1.18.0
